requirements.txt
google-cloud-storage
pycryptodome==3.10.1
